Split cargo/rustc build orchestration out of cli/mod.rs into cargo_build.rs
